implement fetched search results to the general results UI page
