+++ ccpp-jammming/src/components/ui/home/search_container/SearchContainer.scss
