+++ ccpp-jammming/src/components/ui/home/search_container/general_results/GeneralResultsPage.jsx
@@ -4,29 +4,28 @@ import ArtistResults from "./list_components/ArtistResults";
 import AlbumResults from "./list_components/AlbumResults";
 import SongResults from "./list_components/SongResults";
 
-function GeneralResultsPage() {
+function GeneralResultsPage({ songsResults, artistsResults, albumResults }) {
     return (
         <>
             <div id="gen-results-container" className="container-fluid">
                 <div id="gen-results-artists" className="row">
                     <div id="gen-results-artists-col" className="col">
-                        <ArtistResults/>
+                        <ArtistResults artistsResults={artistsResults}/>
                     </div>
                 </div>
-                {/* <div id="gen-results-albuns" className="row">
+                <div id="gen-results-albuns" className="row">
                     <div id="gen-results-albuns-col" className="col">
-                        <AlbumResults/>
+                        <AlbumResults albumResults={albumResults}/>
                     </div>
                 </div>
                 <div id="gen-results-songs" className="row">
                     <div id="gen-results-songs-col" className="col">
-                        <SongResults/>
+                        <SongResults songsResults={songsResults}/>
                     </div>
-                </div> */}
+                </div>
             </div>
         </>
     );       
 }
 
 export default GeneralResultsPage;
-
