+++ ccpp-jammming/src/components/ui/home/search_container/general_results/list_components/ArtistResults.jsx
@@ -2,11 +2,13 @@ import React from "react";
 import IMG from "../../../../../../assets/images/ImagesHUB";
 import Artists from "./unit_components/Artists";
 
-function ArtistResults() {
+function ArtistResults({ artistsResults  }) {
     return (
         <>
             <h4>artists:</h4>
-            <Artists/>
+            {artistsResults.map(artist => (
+                <Artists artist={artist} key={artist.uri}/>
+            ))}
         </>
     );
 }
