+++ ccpp-jammming/src/components/ui/home/search_container/general_results/list_components/SongResults.jsx
@@ -2,11 +2,13 @@ import React from "react";
 import IMG from "../../../../../../assets/images/ImagesHUB";
 import Songs from "./unit_components/Songs";
 
-function SongResults() {
+function SongResults({ songsResults }) {
     return (
         <>
             <h4>songs:</h4>
-            <Songs/>
+            {songsResults.map(song => (
+                <Songs song={song} key={song.uri}/>
+            ))}
         </>
     );
 }
