+++ ccpp-jammming/src/components/ui/home/search_container/general_results/list_components/unit_components/Albuns.jsx
@@ -1,16 +1,16 @@
 import React from "react";
 import IMG from "../../../../../../../assets/images/ImagesHUB";
 
-function Albuns() {
+function Albuns({ album }) {
     return (
         <>
             <div id="albuns-inner-row" className="row">
                 <div id="album-thumbnail" className="col-1 d-flex justify-content-center align-items-center">
-                    <img src={IMG.placeHolders} alt="album cover" height="50px"/>
+                    <img src={album.cover} alt="album cover" height="55px"/>
                 </div>
                 <div id="album-title" className="col d-flex flex-column justify-content-center align-items-start">
-                    <h5>Unlimited Love</h5>
-                    <p>2022 - Red Hot Chili Peppers</p>
+                    <h5>{album.album}</h5>
+                    <p>{album.year} - {album.artist}</p>
                 </div>
             </div>  
         </>
