+++ ccpp-jammming/src/components/ui/home/search_container/general_results/list_components/unit_components/Artists.jsx
@@ -1,15 +1,15 @@
 import React from "react";
 import IMG from "../../../../../../../assets/images/ImagesHUB";
 
-function Artists() {
+function Artists({ artist }) {
     return (
         <>
             <div id="artist-inner-row" className="row">
                 <div id="artist-thumbnail" className="col-1 d-flex justify-content-center align-items-center">
-                    <img src={IMG.profilePlaceHolder} alt="profile picture" height="70px"/>
+                    <img src={artist.profile} alt="profile picture" height="70px"/>
                 </div>
                 <div id="artist-name" className="col d-flex justify-content-start align-items-center">
-                    <h5>Red Hot Chili Peppers</h5>
+                    <h5>{artist.artist}</h5>
                 </div>
             </div>
         </>
