+++ ccpp-jammming/src/components/ui/home/search_container/general_results/list_components/unit_components/Songs.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import IMG from "../../../../../../../assets/images/ImagesHUB";
 
-function Songs() {
+function Songs({ song }) {
     return (
         <>
             <div id="songs-inner-row" className="row">
@@ -9,11 +9,11 @@ function Songs() {
                     <img id="play-icon" src={IMG.play2PNG} alt="play icon" width="22px"/>
                 </div>
                 <div id="col-cover" className="col-1 d-flex justify-content-center align-items-center">
-                    <img src={IMG.placeHolders} height="40px"/>
+                    <img src={song.cover} height="40px"/>
                 </div>
                 <div id="col-title" className="col d-flex justify-content-start align-items-center">
-                    <h5>Aquatic Mouth Dance</h5>
-                    <p>Red Hot Chilli Peppers</p>
+                    <h5>{song.title}</h5>
+                    <p>{song.artist}</p>
                 </div>
                 <div id="col-plus" className="col-1 d-flex justify-content-end align-items-center">
                     <img id="plus-icon" src={IMG.plus2PNG} alt="plus icon" width="25px"/>
@@ -22,10 +22,10 @@ function Songs() {
                     <img id="minus-icon" src={IMG.minus2PNG} alt="minus icon" width="25x"/>
                 </div>
                 <div id="col-album" className="col-2 d-flex justify-content-start align-items-center">
-                    <p>Unlimited Love</p>
+                    <p>{song.album}</p>
                 </div>
                 <div id="col-duration" className="col-1 d-flex justify-content-center align-items-center">
-                    <p>4:20</p>
+                    <p>{song.duration / 60000}</p>
                 </div>
             </div> 
         </>
