+++ ccpp-jammming/src/scss/main.scss
@@ -23,7 +23,7 @@ $form-check-input-focus-box-shadow: none;
 @import "../components/ui/home/playlists_container/open_playlist/OpenPlaylist.scss";
 @import "../components/ui/home/playlists_container/open_playlist/track/Track.scss";
 @import "../components/ui/home/track_player/TrackPlayer.scss";
-@import "../components/ui/home/search_container/SearchBox.scss";
+@import "../components/ui/home/search_container/SearchContainer.scss";
 @import "../components/ui/home/search_container/general_results/GeneralResultsPage.scss";
 @import "../components/ui/home/search_container/general_results/list_components/ListComponents.scss";
 @import "../components/ui/home/search_container/artist_page/ArtistPage.scss";
